Handle pointers to slice of custom types
If a user used a *[]MyType slice in params or responses,
a panic would be raised looking for the data type definition
for *[]MyType. This is because the pointer-to-slice wasn't recognized
as a "thing that should be walked", it was instead recognized
as a "type to export directly".
Fixing this required NewField to take pointer values
and dereference to the underlying type.

However, there are times we don't want to dereference
(like when registering a custom data type definition,
we also want to register the "pointer version" of the type/value).
So I added an internal newField method.

Fixes #4