Fix $ref being used for exported parameter types
If the parameter type had an exported field of an exported type,
it would show up in the requestBody as a $ref,
without a corresponding entry in the schemas portion of the doc.

This change makes it so struct fields will *always* be
expanded for request bodies, even if they are an exported type.
This is not consistent with how response shapes work
($ref is used for exported types, expansion for unexported).
Parameters were set up to be more expansion-heavy from the start,
but we can revisit this if it's a problem in practice
(the swagger is still correct, just more verbose).